daruka-backend/requirements.txt
alembic==1.13.2
anyio==4.10.0
async-timeout==5.0.1
asyncpg==0.29.0
bcrypt==4.3.0
certifi==2025.8.3
click==8.1.8
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.111.0
fastapi-cli==0.0.13
GeoAlchemy2==0.14.6
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.0.2
orjson==3.11.3
packaging==25.0
passlib==1.7.4
pyasn1==0.6.1
pydantic==1.10.14
Pygments==2.19.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
PyYAML==6.0.2
rich==14.1.0
rich-toolkit==0.15.1
rsa==4.9.1
shapely==2.0.4
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
typer==0.19.1
typing_extensions==4.15.0
ujson==5.11.0
uvicorn==0.30.1
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
